# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/trainings": {
      "get": {
        "operationId": "TrainingController_listTrainings",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Lista de treinamentos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TrainingListItemView"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar todos os treinamentos",
        "tags": [
          "Trainings"
        ]
      },
      "post": {
        "operationId": "TrainingController_createTraining",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTrainingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Treinamento criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Criar novo treinamento",
        "tags": [
          "Trainings"
        ]
      }
    },
    "/trainings/{id}": {
      "put": {
        "operationId": "TrainingController_updateTraining",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTrainingDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Treinamento atualizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Atualizar treinamento",
        "tags": [
          "Trainings"
        ]
      },
      "delete": {
        "operationId": "TrainingController_deleteTraining",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Treinamento deletado com sucesso"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Deletar treinamento",
        "tags": [
          "Trainings"
        ]
      }
    },
...
    "/dependants": {
      "post": {
        "operationId": "DependantController_add",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddDependantDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Dependente adicionado com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DependantDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Adiciona um novo dependente à família do usuário",
        "tags": [
          "Família e Dependentes"
        ]
      },
      "get": {
        "operationId": "DependantController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Lista de dependentes retornada com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DependantDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Lista todos os dependentes da família do usuário",
        "tags": [
          "Família e Dependentes"
        ]
      }
    },
    "/dependants/my-family": {
      "get": {
        "operationId": "DependantController_viewMyFamily",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Dados da família retornados com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FamilyDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Visualiza os dados da família do usuário autenticado",
        "tags": [
          "Família e Dependentes"
        ]
      }
    },
    "/dependants/{id}": {
      "get": {
        "operationId": "DependantController_viewDependant",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Dados de um dependente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ViewDependantOutputDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Visualiza os dados de um dependente.",
        "tags": [
          "Família e Dependentes"
        ]
      },
      "patch": {
        "operationId": "DependantController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDependantDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Dependente atualizado com sucesso."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Atualiza os dados de um dependente específico",
        "tags": [
          "Família e Dependentes"
        ]
      },
      "delete": {
        "operationId": "DependantController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Dependente removido com sucesso."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Remove um dependente da família",
        "tags": [
          "Família e Dependentes"
        ]
      }
    },
...
  },
  "info": {
    "title": "API NCFCA",
    "description": "Documentação da API para o sistema NCFCA. Esta é a única fonte de verdade para os contratos da API.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Entre com o token JWT",
        "in": "header"
      }
    },
    "schemas": {
...
      "DependantDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID único do dependente.",
            "format": "uuid",
            "example": "a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11"
          },
          "firstName": {
            "type": "string",
            "description": "Primeiro nome do dependente.",
            "example": "João"
          },
          "lastName": {
            "type": "string",
            "description": "Último nome do dependente.",
            "example": "Silva Filho"
          },
          "birthdate": {
            "type": "string",
            "description": "Data de nascimento do dependente.",
            "format": "date-time",
            "example": "2010-01-20T00:00:00.000Z"
          },
          "relationship": {
            "type": "string",
            "description": "Relação de parentesco com o responsável.",
            "enum": [
              "DAUGHTER",
              "HUSBAND",
              "CHILD",
              "WIFE",
              "SON",
              "OTHER"
            ],
            "example": "SON"
          },
          "sex": {
            "type": "string",
            "description": "Sexo do dependente.",
            "enum": [
              "FEMALE",
              "MALE"
            ],
            "example": "MALE"
          },
          "type": {
            "type": "string",
            "description": "Tipo do dependente.",
            "enum": [
              "STUDENT",
              "PARENT",
              "ALUMNI"
            ],
            "example": "STUDENT"
          },
          "email": {
            "type": "object",
            "description": "Email de contato do dependente (opcional).",
            "example": "joao.filho@example.com"
          },
          "phone": {
            "type": "object",
            "description": "Telefone de contato do dependente (opcional).",
            "example": "11999998888"
          },
          "familyId": {
            "type": "string",
            "description": "ID da família à qual o dependente pertence.",
            "format": "uuid",
            "example": "f1eebc99-9c0b-4ef8-bb6d-6bb9bd380a22"
          }
        },
        "required": [
          "id",
          "firstName",
          "lastName",
          "birthdate",
          "relationship",
          "sex",
          "type",
          "familyId"
        ]
      },
      "FamilyDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID único da família.",
            "format": "uuid",
            "example": "a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11"
          },
          "holderId": {
            "type": "string",
            "description": "ID do usuário responsável pela família (holder).",
            "format": "uuid",
            "example": "f1eebc99-9c0b-4ef8-bb6d-6bb9bd380a22"
          },
          "status": {
            "type": "string",
            "description": "Status atual da afiliação da família.",
            "enum": [
              "PENDING_PAYMENT",
              "NOT_AFFILIATED",
              "AFFILIATED",
              "EXPIRED"
            ],
            "example": "AFFILIATED"
          },
          "affiliatedAt": {
            "type": "string",
            "description": "Data em que a afiliação se tornou ativa.",
            "format": "date-time",
            "example": "2025-07-17T03:00:00.000Z",
            "nullable": true
          },
          "affiliationExpiresAt": {
            "type": "string",
            "description": "Data em que a afiliação irá expirar.",
            "format": "date-time",
            "example": "2026-07-17T03:00:00.000Z",
            "nullable": true
          },
          "dependants": {
            "description": "Lista de dependentes associados a esta família.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DependantDto"
            }
          }
        },
        "required": [
          "id",
          "holderId",
          "status",
          "affiliatedAt",
          "affiliationExpiresAt",
          "dependants"
        ]
      },
...
      "TrainingListItemView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID único do treinamento",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "description": "Título do treinamento",
            "example": "Como fazer passes básicos"
          },
          "description": {
            "type": "string",
            "description": "Descrição do treinamento",
            "example": "Aprenda os fundamentos dos passes no futebol"
          },
          "youtubeUrl": {
            "type": "string",
            "description": "URL do vídeo no YouTube",
            "example": "https://youtube.com/watch?v=abc123"
          }
        },
        "required": [
          "id",
          "title",
          "description",
          "youtubeUrl"
        ]
      },
      "CreateTrainingDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Título do treinamento",
            "example": "Como fazer passes básicos"
          },
          "description": {
            "type": "string",
            "description": "Descrição do treinamento",
            "example": "Aprenda os fundamentos dos passes no futebol"
          },
          "youtubeUrl": {
            "type": "string",
            "description": "URL do vídeo no YouTube",
            "example": "https://youtube.com/watch?v=abc123"
          }
        },
        "required": [
          "title",
          "description",
          "youtubeUrl"
        ]
      },
      "TrainingResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID único do treinamento",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "description": "Título do treinamento"
          },
          "description": {
            "type": "string",
            "description": "Descrição do treinamento"
          },
          "youtubeUrl": {
            "type": "string",
            "description": "URL do vídeo no YouTube"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Data de criação"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Data de atualização"
          }
        },
        "required": [
          "id",
          "title",
          "description",
          "youtubeUrl",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateTrainingDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Título do treinamento",
            "example": "Como fazer passes básicos"
          },
          "description": {
            "type": "string",
            "description": "Descrição do treinamento",
            "example": "Aprenda os fundamentos dos passes no futebol"
          },
          "youtubeUrl": {
            "type": "string",
            "description": "URL do vídeo no YouTube",
            "example": "https://youtube.com/watch?v=abc123"
          }
        },
        "required": [
          "title",
          "description",
          "youtubeUrl"
        ]
      },
...
      "AddDependantDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "description": "Primeiro nome do dependente.",
            "example": "João",
            "minLength": 2
          },
          "lastName": {
            "type": "string",
            "description": "Último nome do dependente.",
            "example": "Silva",
            "minLength": 2
          },
          "birthdate": {
            "type": "string",
            "description": "Data de nascimento do dependente no formato AAAA-MM-DD.",
            "example": "2010-05-20"
          },
          "relationship": {
            "type": "string",
            "description": "Relação de parentesco com o responsável familiar.",
            "enum": [
              "DAUGHTER",
              "HUSBAND",
              "CHILD",
              "WIFE",
              "SON",
              "OTHER"
            ],
            "example": "SON"
          },
          "sex": {
            "type": "string",
            "description": "Sexo do dependente.",
            "enum": [
              "FEMALE",
              "MALE"
            ],
            "example": "MALE"
          },
          "email": {
            "type": "string",
            "description": "Email de contato do dependente.",
            "example": "joao.silva@example.com"
          },
          "phone": {
            "type": "string",
            "description": "Telefone de contato do dependente (opcional).",
            "example": "11987654321"
          }
        },
        "required": [
          "firstName",
          "lastName",
          "birthdate",
          "relationship",
          "sex",
          "email"
        ]
      },
      "HolderDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID único do titular.",
            "format": "uuid",
            "example": "b1eebc99-9c0b-4ef8-bb6d-6bb9bd380a22"
          },
          "firstName": {
            "type": "string",
            "description": "Nome do titular.",
            "example": "Maria Silva"
          },
          "lastName": {
            "type": "string",
            "description": "Nome do titular.",
            "example": "Maria Silva"
          },
          "email": {
            "type": "string",
            "description": "Email do titular.",
            "example": "maria.silva@email.com"
          },
          "phone": {
            "type": "string",
            "description": "Telefone do titular.",
            "example": "(11) 98888-8888"
          }
        },
        "required": [
          "id",
          "firstName",
          "lastName",
          "email",
          "phone"
        ]
      },
      "ViewDependantOutputDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID único do dependente.",
            "format": "uuid",
            "example": "a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11"
          },
          "firstName": {
            "type": "string",
            "description": "Primeiro nome do dependente.",
            "example": "João"
          },
          "lastName": {
            "type": "string",
            "description": "Último nome do dependente.",
            "example": "Silva Filho"
          },
          "birthdate": {
            "type": "string",
            "description": "Data de nascimento do dependente.",
            "format": "date-time",
            "example": "2010-01-20T00:00:00.000Z"
          },
          "relationship": {
            "type": "string",
            "description": "Relação de parentesco com o responsável.",
            "enum": [
              "DAUGHTER",
              "HUSBAND",
              "CHILD",
              "WIFE",
              "SON",
              "OTHER"
            ],
            "example": "SON"
          },
          "sex": {
            "type": "string",
            "description": "Sexo do dependente.",
            "enum": [
              "FEMALE",
              "MALE"
            ],
            "example": "MALE"
          },
          "email": {
            "type": "object",
            "description": "Email de contato do dependente (opcional).",
            "example": "joao.filho@example.com"
          },
          "phone": {
            "type": "object",
            "description": "Telefone de contato do dependente (opcional).",
            "example": "11999998888"
          },
          "familyId": {
            "type": "string",
            "description": "ID da família à qual o dependente pertence.",
            "format": "uuid",
            "example": "f1eebc99-9c0b-4ef8-bb6d-6bb9bd380a22"
          },
          "holder": {
            "description": "Dados do titular.",
            "allOf": [
              {
                "$ref": "#/components/schemas/HolderDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "firstName",
          "lastName",
          "birthdate",
          "relationship",
          "sex",
          "familyId",
          "holder"
        ]
      },
      "UpdateDependantDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "description": "Novo primeiro nome do dependente.",
            "example": "Joana",
            "minLength": 2
          },
          "lastName": {
            "type": "string",
            "description": "Novo último nome do dependente.",
            "example": "Souza",
            "minLength": 2
          },
          "birthdate": {
            "type": "string",
            "description": "Nova data de nascimento do dependente (AAAA-MM-DD).",
            "example": "2011-03-15"
          },
          "relationship": {
            "type": "string",
            "description": "Nova relação de parentesco.",
            "enum": [
              "DAUGHTER",
              "HUSBAND",
              "CHILD",
              "WIFE",
              "SON",
              "OTHER"
            ],
            "example": "DAUGHTER"
          },
          "sex": {
            "type": "string",
            "description": "Novo sexo do dependente.",
            "enum": [
              "FEMALE",
              "MALE"
            ],
            "example": "FEMALE"
          },
          "email": {
            "type": "string",
            "description": "Novo email do dependente.",
            "example": "joana.souza@example.com"
          },
          "phone": {
            "type": "string",
            "description": "Novo telefone do dependente.",
            "example": "11912345678"
          }
        }
      },
...
    }
  }
}